Remove all warnings and treat them as build errors
